Modify simple_html_dom to fit our style
